+++ test/percentile_test.rb
@@ -6,14 +6,11 @@ class PercentileTest < Minitest::Test
   end
 
   def test_even
-    skip if sqlite?
     [1, 2, 3, 4].each { |n| User.create!(visits_count: n) }
     assert_in_delta 3.25, User.percentile(:visits_count, 0.75)
   end
 
   def test_odd
-    skip if sqlite?
-
     [15, 20, 35, 40, 50].each { |n| User.create!(visits_count: n) }
     assert_equal 29, User.percentile(:visits_count, 0.4)
   end
@@ -29,8 +26,6 @@ class PercentileTest < Minitest::Test
   end
 
   def test_high
-    skip if sqlite?
-
     [1, 1, 2, 3, 4, 100].each { |n| User.create!(visits_count: n) }
     assert_in_delta 95.2, User.percentile(:visits_count, 0.99)
   end
@@ -54,14 +49,14 @@ class PercentileTest < Minitest::Test
   end
 
   def test_expression
-    skip if mongoid? || sqlite?
+    skip if mongoid?
 
     [1, 2, 3, 4].each { |n| User.create!(visits_count: n) }
     assert_in_delta 4.25, User.percentile(Arel.sql("visits_count + 1"), 0.75)
   end
 
   def test_expression_no_arel
-    skip if mongoid? || sqlite?
+    skip if mongoid?
 
     error = assert_raises(ActiveRecord::UnknownAttributeReference) do
       User.percentile("visits_count + 1", 0.75)